pyproject.toml
[tool.poetry]
name = "mcp-audio-server"
version = "0.1.0"
description = "MCP audio processing server"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.27.0"
pydantic = "^2.6.0"
numpy = "^1.26.0"
soundfile = "^0.12.1"
jsonschema = "^4.20.0"
structlog = "^23.2.0"
python-json-logger = "^2.0.7"
uuid = "^1.30"
librosa = "^0.10.1"
prometheus-client = "^0.19.0"
psutil = "^5.9.8"
python-magic = "^0.4.27"
redis = "^5.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
